extrated the difficultyScore to a seperate class (Difficultycontroller). Also fixed the issue where the difficulty score is increasing when falling from the one leg track.
