--- FYP/Assets/AudioManager.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: b368b3ed646464043a06a0ab7b22abed
-timeCreated: 1523434366
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
