--- FYP/Library/ScriptAssemblies/BuiltinAssemblies.stamp
@@ -1,2 +1,0 @@
-0000.5a2fe4d8.0000
-0000.5a2ff7a2.0000
